fix: Accordion overlay content [MDS-352]
