+++ workspaces/core/src/accordion/private/types/ContentProps.ts
@@ -1,0 +1,6 @@
+type ContentProps = {
+  className?: string;
+  children?: React.ReactNode;
+};
+
+export default ContentProps;
