+++ workspaces/core/src/accordion/private/utils/AccordionItemContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type AccordionItemState from '../types/AccordionItemState';
+
+const AccordionItemContext = createContext<AccordionItemState>({});
+AccordionItemContext.displayName = 'AccordionItemContext';
+
+export default AccordionItemContext;
